pyproject.toml
[project]
name = "travel-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Timothy Wong",email = "timothy.wong@hotmail.co.uk"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "openai (>=1.93.0,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langgraph (>=0.5.1,<0.6.0)",
    "langchain-openai (>=0.3.27,<0.4.0)",
    "langgraph-supervisor (>=0.0.27,<0.0.28)",
    "langchain-chroma (>=0.2.4,<0.3.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pytest = "^8.4.1"
pytest-mock = "^3.14.1"

